# backend/services/SSEService/openapi.yaml
openapi: 3.0.3
info:
  title: Knuffel SSE Service API
  description: |
    Server-Sent Events (SSE) service for the Knuffel multiplayer Kniffel application (MVP).

    Responsibilities:
    - SSE subscription for lobby audiences only (no game stream)
    - Central broadcasting of events published by internal services
    - Connection lifecycle management with per-connection keep-alives

    Behavior:
    - Single public stream endpoint: GET /events/lobby/{lobby_id}
    - Addressing strictly via lobby_id (UUID v4)
    - JWT validation by SSEService directly against AuthService via JWTCookie on the subscription endpoint
    - Membership check via LobbyService internal endpoint
    - event_type is free-form (non-empty, max 128 chars); examples include lobby_updated and game_updated; keep_alive is reserved and emitted only by SSEService
    - data is always a JSON object; timestamp is injected/overwritten by SSEService as Unix epoch milliseconds (number) with field name "timestamp"
    - Keep-alive emitted every 30s per connection; no retry directive line is sent
    - Rate-limits disabled in MVP for SSE streams
    - Internal endpoints are isolated via reverse proxy; no auth configured in this spec for /internal/*
    - Registry emerges/disappears implicitly via subscriptions (no explicit register/unregister)

  version: 1.0.0
  contact:
    name: Knuffel Team

servers:
  - url: http://localhost:8084
    description: SSEService (MVP)

tags:
  - name: SSE Streams
    description: Client SSE connection endpoints
  - name: Internal
    description: Internal event publishing endpoints

paths:
  /events/lobby/{lobby_id}:
    get:
      tags:
        - SSE Streams
      summary: Subscribe to lobby events
      description: |
        Opens a Server-Sent Events stream for lobby updates. Connection stays open until client closes or server terminates.

        Rules:
        - event_type is free-form (non-empty, ≤128 chars). Commonly used: lobby_updated, game_updated. keep_alive is reserved and emitted by the service every 30s per connection.
        - data is a JSON object (always). The service injects/overwrites data.timestamp with the current server time as Unix epoch milliseconds (number).
        - Membership is verified against LobbyService. If lobby is not found → 404; if the user is not a member → 403.

        Authentication:
        - Requires JWT token in cookie (JWTCookie) on this endpoint.
      operationId: subscribeLobbyEvents
      parameters:
        - name: lobby_id
          in: path
          required: true
          description: Lobby identifier (UUID v4)
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: jwt
          in: cookie
          required: true
          description: JWT authentication token
          schema:
            type: string
      security:
        - JWTCookie: []
      responses:
        '200':
          description: SSE stream established
          content:
            text/event-stream:
              schema:
                type: string
                description: Server-Sent Events stream
              examples:
                lobbyUpdated:
                  summary: Lobby updated event
                  value: |
                    event: lobby_updated
                    data: {"timestamp": 1733227200000, "players": [{"user_id":"usr_alice123"}]}
                gameUpdated:
                  summary: Game updated event
                  value: |
                    event: game_updated
                    data: {"timestamp": 1733227200000, "state": {"turn":"usr_abc123"}}
                keepAlive:
                  summary: Keep-alive heartbeat
                  value: |
                    event: keep_alive
                    data: {"timestamp": 1733227200000}
        '401':
          description: Invalid or expired JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidToken:
                  summary: Invalid or expired JWT
                  value:
                    error: "unauthorized"
                    message: "Invalid or expired authentication token"
        '403':
          description: Not a member of this lobby
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notInLobby:
                  summary: User not in lobby
                  value:
                    error: "forbidden"
                    message: "You are not a member of this lobby"
        '404':
          description: Lobby not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                lobbyNotFound:
                  summary: Lobby does not exist
                  value:
                    error: "lobby_not_found"
                    message: "Lobby not found"

  /internal/publish:
    post:
      tags:
        - Internal
      summary: Publish event to lobby connections
      description: |
        Publishes an event to all SSE connections for a specific lobby.
        Only accessible from other internal services (e.g., LobbyService, GameService) via reverse proxy.

        Behavior:
        - SSEService injects/overwrites data.timestamp with server time in epoch ms (number).
        - If data is absent, SSEService uses {"timestamp": <epoch_ms>}.
        - If event_type == "keep_alive" → 400 invalid_request.
        - If data is provided but is not an object → 400 invalid_request.
        - Publishing to unknown lobby_id in registry → 404 target_not_found.
        - If no active connections exist for the lobby → 200 success=true with zero counters.
      operationId: publishEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublishEventRequest'
            examples:
              lobbyUpdated:
                summary: Publish lobby updated
                value:
                  lobby_id: "550e8400-e29b-41d4-a716-446655440000"
                  event_type: "lobby_updated"
                  data:
                    timestamp: 1733227200000
                    players:
                      - user_id: "usr_alice123"
                      - user_id: "usr_bob456"
              gameUpdated:
                summary: Publish game updated for lobby audience
                value:
                  lobby_id: "550e8400-e29b-41d4-a716-446655440000"
                  event_type: "game_updated"
                  data:
                    timestamp: 1733227200000
                    state:
                      turn: "usr_abc123"
      responses:
        '200':
          description: Event published successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublishEventResponse'
              examples:
                success:
                  summary: Event delivered to all connections
                  value:
                    success: true
                    connections_found: 3
                    events_sent: 3
                    failed_connections: 0
                partialSuccess:
                  summary: Some connections failed
                  value:
                    success: true
                    connections_found: 3
                    events_sent: 2
                    failed_connections: 1
                noConnections:
                  summary: No active connections
                  value:
                    success: true
                    connections_found: 0
                    events_sent: 0
                    failed_connections: 0
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Target lobby not found in connection registry
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                notFound:
                  summary: Target lobby not found
                  value:
                    error: "target_not_found"
                    message: "Target lobby not found"
        '500':
          $ref: '#/components/responses/InternalServerError'

  /healthcheck:
    get:
      tags:
        - Internal
      summary: Health check endpoint
      description: Returns service health status
      operationId: healthCheck
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
              examples:
                healthy:
                  summary: Healthy service
                  value:
                    status: "ok"

components:
  schemas:
    PublishEventRequest:
      type: object
      required:
        - lobby_id
        - event_type
      properties:
        lobby_id:
          type: string
          format: uuid
          description: Lobby identifier (UUID v4)
          example: "550e8400-e29b-41d4-a716-446655440000"
        event_type:
          type: string
          minLength: 1
          maxLength: 128
          description: >
            Event type name (sent as SSE event field). Free-form; "keep_alive" is reserved and will be rejected with 400 invalid_request.
          example: "lobby_updated"
        data:
          type: object
          nullable: true
          additionalProperties: true
          description: >
            Optional payload object (sent as SSE data field as JSON). If provided, must be a JSON object.
            SSEService injects/overwrites data.timestamp with server time in Unix epoch milliseconds (number).
          example:
            timestamp: 1733227200000
            any: "value"

    PublishEventResponse:
      type: object
      required:
        - success
        - connections_found
        - events_sent
        - failed_connections
      properties:
        success:
          type: boolean
          description: Whether operation completed
          example: true
        connections_found:
          type: integer
          description: Number of connections found for target
          minimum: 0
          example: 3
        events_sent:
          type: integer
          description: Number of events successfully sent
          minimum: 0
          example: 3
        failed_connections:
          type: integer
          description: Number of connections that failed
          minimum: 0
          example: 0

    HealthResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum:
            - ok
          description: Service health status
          example: "ok"

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "target_not_found"
        message:
          type: string
          description: Human-readable error message
          example: "No active connections for target"
        details:
          type: object
          description: Additional error details
          additionalProperties: true

  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingField:
              summary: Missing required field
              value:
                error: "invalid_request"
                message: "Missing required field: lobby_id"
            invalidEventTypeKeepAlive:
              summary: event_type keep_alive is reserved
              value:
                error: "invalid_request"
                message: "event_type 'keep_alive' is reserved and cannot be published"
            invalidDataType:
              summary: data must be an object when provided
              value:
                error: "invalid_request"
                message: "data must be a JSON object if provided"

    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            generic:
              summary: Generic server error
              value:
                error: "internal_error"
                message: "An unexpected error occurred"

  securitySchemes:
    JWTCookie:
      type: apiKey
      in: cookie
      name: jwt
      description: JWT authentication token (HTTP-Only cookie)